reorganized folder structure to clients and added cloudinaryclient and supabaseclient in it
